refactor - moved cli to its own dir and removed init from package.json exports
