+++ packages/client/package.json
@@ -7,12 +7,8 @@
       "types": "./dist/client/index.d.ts"
     },
     "./cli": {
-      "import": "./dist/client/cli.js",
-      "types": "./dist/client/cli.d.ts"
-    },
-    "./init": {
-      "import": "./dist/client/init.js",
-      "types": "./dist/client/init.d.ts"
+      "import": "./dist/cli/index.js",
+      "types": "./dist/cli/index.d.ts"
     },
     "./types": {
       "import": "./dist/types/index.js",
