+++ packages/client/src/cli/index.ts
@@ -1,0 +1,5 @@
+export {
+  initCli,
+  passTime,
+} from "./cli.js"
+
