+++ packages/client/src/client/index.ts
@@ -3,6 +3,11 @@ export {
 } from "./setup.js"
 
 export {
+  connect,
+  getNewToken,
+} from "./init.js"
+
+export {
   type ClientSockets,
   type SetupSocketOptions
 } from "./setup.js"
