+++ packages/client/src/client/init.ts
@@ -1,28 +1,7 @@
 import { GameClient } from "./game/gameclient.js";
 import { SocketManager } from "./sockets/socketManager.js";
-import { getAuthToken, AuthRequestData } from "./../utils.js";
+import { getAuthToken } from "./../utils.js";
 
-export async function initClient(name: string, url: string, gameId: string) {
-
-  const data: AuthRequestData = { name: name, url: url };
-  const newToken = await getAuthToken(data);
-
-  if (newToken) {
-    const sockets = new SocketManager({
-      serverUrl: url,
-      token: newToken,
-      gameId: gameId, 
-    });
-
-    await sockets.socketsReady();
-
-    const client = new GameClient(sockets.token, sockets);
-
-    await client.play();
-  } else {
-    console.log("Could not get user token from gamemaster in /auth")
-  } 
-}
 
 export async function getNewToken(name: string, url: string) { 
   const token = await getAuthToken({name, url});
