+++ packages/frontend/package.json
@@ -13,7 +13,7 @@
     "format": "prettier --write src/**/*.{ts,html,css}"
   },
   "devDependencies": {
-    "esbuild": "^0.25.2",
+    "esbuild": "^0.25.3",
     "esbuild-serve": "^1.0.1",
     "serve": "^14.2.4"
   },
